Revert to android plugin 3.1.1 Set RPM to integer

+++ app/src/main/java/io/github/thymythos/diagnosticdataviewer/AFRFragment.java
@@ -27,7 +27,7 @@ public class AFRFragment extends Fragment implements LiveDataFragment {
     }
 
     @Override
-    public void setRPM(float rpm) {
+    public void setRPM(int rpm) {
     }
 
     @Override
